Unified embedding tests
